+++ projects/GKv2/GEDmill/GEDmillPlugin.cs
@@ -1,0 +1,117 @@
+﻿/*
+ *  "GEDKeeper", the personal genealogical database editor.
+ *  Copyright (C) 2019 by Sergey V. Zhdanovskih.
+ *
+ *  This file is part of "GEDKeeper".
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+using System;
+using GKCore;
+using GKCore.Interfaces;
+using GKCore.Plugins;
+
+namespace GEDmill
+{
+    /// <summary>
+    /// Plugin's Localizable strings (PLS)
+    /// </summary>
+    public enum PLS
+    {
+        LSID_Title,
+        LSID_1,
+        LSID_2,
+        LSID_3,
+        LSID_4,
+        LSID_5
+    }
+
+    public sealed class Plugin : WidgetPlugin
+    {
+        private string fDisplayName = "Website Generator (GEDmill)";
+        private ILangMan fLangMan;
+
+        public override string DisplayName { get { return fDisplayName; } }
+        public override ILangMan LangMan { get { return fLangMan; } }
+        public override IImage Icon { get { return null; } }
+        public override PluginCategory Category { get { return PluginCategory.Report; } }
+
+        private MainForm fForm;
+
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing) {
+                CloseForm();
+            }
+            base.Dispose(disposing);
+        }
+
+        internal void CloseForm()
+        {
+            if (fForm != null) {
+                fForm = null;
+            }
+        }
+
+        public override void Execute()
+        {
+            if (!Host.IsWidgetActive(this)) {
+                fForm = new MainForm(this);
+                fForm.ShowDialog(); // FIXME
+            } else {
+                fForm.Close();
+            }
+        }
+
+        public override void OnLanguageChange()
+        {
+            try {
+                fLangMan = Host.CreateLangMan(this);
+                fDisplayName = fLangMan.LS(PLS.LSID_Title);
+
+                if (fForm != null)
+                    fForm.SetLang();
+            } catch (Exception ex) {
+                Logger.LogWrite("GEDmillPlugin.OnLanguageChange(): " + ex.Message);
+            }
+        }
+
+        public override bool Shutdown()
+        {
+            bool result = true;
+            try {
+                CloseForm();
+            } catch (Exception ex) {
+                Logger.LogWrite("GEDmillPlugin.Shutdown(): " + ex.Message);
+                result = false;
+            }
+            return result;
+        }
+
+        public override void BaseChanged(IBaseWindow baseWin)
+        {
+            if (fForm != null) {
+                fForm.BaseChanged(baseWin);
+            }
+        }
+
+        public override void BaseClosed(IBaseWindow baseWin)
+        {
+            if (fForm != null) {
+                fForm.BaseChanged(null);
+            }
+        }
+    }
+}