+++ projects/GKv2/GEDmill/CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## 11.08.2019
+- Convert project GEDmill to plugin.
+
+## 10.08.2019
+- Merger of GEDKeeper and GEDmill projects.
+
 ## 30.03.2019
 - Replacing the base on GKCore; reformatting, partial cleaning.